Actualizado modelo
Quitado horasconvenio de las herencias y puesto en trabajador, cambiado
ausenciastrabajador (nombre) y eliminado convenioAusencias del proyecto
y sustituido por ausenciastrabajador
