+++ Controlador/Administracion/Controlador.php
@@ -16,7 +16,7 @@ use Modelo\Base\ParteProducionTarea;
 use Modelo\Base\Produccion;
 use Modelo\Base\TiposFranjas;
 use Modelo\Base\Trabajador;
-use Modelo\Base\TrabajadorAusencia;
+use Modelo\Base\AusenciaTrabajador;
 use Modelo\Base\Vehiculo;
 use Modelo\Base\Viaje;
 use Modelo\BD;
