+++ Modelo/Base/LogisticaClass.php
@@ -20,21 +20,17 @@ require_once __DIR__."/../BD/HorasConvenioBD.php";
 class Logistica extends Trabajador{
 
     private $parteLogistica;
-    private $horasConvenio;
 
     /**
      * Logistica constructor.
      * @param $parteLogistica
      */
-    public function __construct($dni = null, $nombre = null, $apellido1 = null, $apellido2 = null, $telefono = null, $foto = null, $centro = null,  $trabajadorAusencias = null, $horariosTrabajador = null,$parteLogistica=null,$horasConvenio=null)
+    public function __construct($dni = null, $nombre = null, $apellido1 = null, $apellido2 = null, $telefono = null, $foto = null, $centro = null,  $ausenciasTrabajador = null, $horariosTrabajador = null,$parteLogistica=null,$horasConvenio=null)
     {
-        parent::__construct($dni, $nombre, $apellido1, $apellido2, $telefono, $foto, $centro,  $trabajadorAusencias, $horariosTrabajador);
+        parent::__construct($dni , $nombre , $apellido1 , $apellido2 , $telefono , $foto , $centro ,  $ausenciasTrabajador , $horariosTrabajador, $horasConvenio);
         $this->setParteLogistica($parteLogistica);
-        $this->setHorasConvenio($horasConvenio);
 
     }
-
-
     /**
      * @return mixed
      */
@@ -53,29 +49,5 @@ class Logistica extends Trabajador{
     {
         $this->parteLogistica = $parteLogistica;
     }
-
-    /**
-     * @return mixed
-     */
-    public function getHorasConvenio()
-    {
-        if($this->horasConvenio==null){
-            $this->setHorasConvenio(BD\HorasConvenioBD::getHorasConvenioByPerfil($this));
-        }
-        return $this->horasConvenio;
-    }
-
-    /**
-     * @param mixed $horasConvenio
-     */
-    public function setHorasConvenio($horasConvenio)
-    {
-        $this->horasConvenio = $horasConvenio;
-    }
-
-
-
-
-
 
 }
